requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.3
djangorestframework==3.13.1
flake8==4.0.1
isort==5.10.1
kombu==5.2.4
mccabe==0.7.0
packaging==21.3
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
